[UPD]  (BROKE CMAKE) adding basic rooms
