+++ client/publicInclude/Client/Client.hpp
@@ -1,5 +1,5 @@
 #pragma once
-#include "../../Include/ClientData.hpp"
+// #include "../../Include/ClientData.hpp"
 #include "../../Include/ClientServer.hpp"
-#include "../../Include/TCPClient.hpp"
-#include "../../Include/UDPClient.hpp"
+// #include "../../Include/TCPClient.hpp"
+// #include "../../Include/UDPClient.hpp"
